everything edited arpit

+++ base/com/controller/city_controller.py
@@ -4,15 +4,18 @@ from base import app
 from base.com.dao.city_dao import CityDAO
 from base.com.dao.state_dao import StateDAO
 from base.com.vo.city_vo import CityVO
+from base.com.controller.login_controller import admin_login_session
 
 
 @app.route('/admin/load_city', methods=['get'])
 def admin_load_city():
     try:
-        state_dao = StateDAO()
-        state_vo_list = state_dao.view_state()
-        return render_template('admin/addCity.html', state_vo_list=state_vo_list)
-
+        if admin_login_session() == 'admin':
+            state_dao = StateDAO()
+            state_vo_list = state_dao.view_state()
+            return render_template('admin/addCity.html', state_vo_list=state_vo_list)
+        else:
+            return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print('admin_view_city route exception occured>>>>>>>>>>', ex)
 
@@ -20,15 +23,19 @@ def admin_load_city():
 @app.route('/admin/insert_city', methods=['post'])
 def admin_add_city():
     try:
-        city_vo = CityVO()
-        city_dao = CityDAO()
+        if admin_login_session() == 'admin':
 
-        city_vo.city_name = request.form.get('cityName')
-        city_vo.city_description = request.form.get('cityDescription')
-        city_vo.city_state_id = request.form.get('stateId')
+            city_vo = CityVO()
+            city_dao = CityDAO()
 
-        city_dao.insert_city(city_vo)
-        return redirect(url_for('admin_view_city'))
+            city_vo.city_name = request.form.get('cityName')
+            city_vo.city_description = request.form.get('cityDescription')
+            city_vo.city_state_id = request.form.get('stateId')
+
+            city_dao.insert_city(city_vo)
+            return redirect(url_for('admin_view_city'))
+        else:
+             return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_insert_city route exception occured>>>>>>>>>>", ex)
@@ -37,10 +44,12 @@ def admin_add_city():
 @app.route('/admin/view_city', methods=['get'])
 def admin_view_city():
     try:
-        city_dao = CityDAO()
-        city_vo_list = city_dao.view_city()
-        return render_template('admin/viewCity.html', city_vo_list=city_vo_list)
-
+        if admin_login_session() == 'admin':
+            city_dao = CityDAO()
+            city_vo_list = city_dao.view_city()
+            return render_template('admin/viewCity.html', city_vo_list=city_vo_list)
+        else:
+             return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print("admin_view_city route exception occured>>>>>>>>>>", ex)
 
@@ -48,10 +57,13 @@ def admin_view_city():
 @app.route('/admin/delete_city', methods=['get'])
 def admin_delete_city():
     try:
-        city_dao = CityDAO()
-        city_id = request.args.get('cityId')
-        city_dao.delete_city(city_id)
-        return redirect(url_for('admin_view_city'))
+        if admin_login_session() == 'admin':
+            city_dao = CityDAO()
+            city_id = request.args.get('cityId')
+            city_dao.delete_city(city_id)
+            return redirect(url_for('admin_view_city'))
+        else:
+             return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print("in admin_delete_city route exception occured>>>>>>>>>>", ex)
 
@@ -59,15 +71,18 @@ def admin_delete_city():
 @app.route('/admin/edit_city', methods=['get'])
 def admin_edit_city():
     try:
-        city_vo = CityVO()
-        city_dao = CityDAO()
-        state_dao = StateDAO()
+        if admin_login_session() == 'admin':
+            city_vo = CityVO()
+            city_dao = CityDAO()
+            state_dao = StateDAO()
 
-        city_vo.city_id = request.args.get('cityId')
-        city_vo_list = city_dao.edit_city(city_vo)
-        state_vo_list = state_dao.view_state()
-        return render_template('admin/editCity.html', state_vo_list=state_vo_list,
-                               city_vo_list=city_vo_list)
+            city_vo.city_id = request.args.get('cityId')
+            city_vo_list = city_dao.edit_city(city_vo)
+            state_vo_list = state_dao.view_state()
+            return render_template('admin/editCity.html', state_vo_list=state_vo_list,
+                                   city_vo_list=city_vo_list)
+        else:
+             return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print("in admin_edit_city route exception occured>>>>>>>>>>", ex)
 
@@ -75,14 +90,17 @@ def admin_edit_city():
 @app.route('/admin/update_city', methods=['POST'])
 def admin_update_city():
     try:
-        city_vo = CityVO()
-        city_dao = CityDAO()
+        if admin_login_session() == 'admin':
+            city_vo = CityVO()
+            city_dao = CityDAO()
 
-        city_vo.city_id = request.form.get('cityId')
-        city_vo.city_name = request.form.get('cityName')
-        city_vo.city_description = request.form.get('cityDescription')
-        city_vo.city_state_id = request.form.get('cityStateId')
-        city_dao.update_city(city_vo)
-        return redirect(url_for('admin_view_city'))
+            city_vo.city_id = request.form.get('cityId')
+            city_vo.city_name = request.form.get('cityName')
+            city_vo.city_description = request.form.get('cityDescription')
+            city_vo.city_state_id = request.form.get('cityStateId')
+            city_dao.update_city(city_vo)
+            return redirect(url_for('admin_view_city'))
+        else:
+             return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print("in admin_update_city route exception occured>>>>>>>>>>", ex)
